Split ACT function of mel.go into smaller helper functions
